+++ controllers/TestdataController.php
@@ -32,13 +32,19 @@ class TestdataController extends Controller
      */
     public function actionIndex()
     {
-        $dataProvider = new ActiveDataProvider([
-            'query' => Test::find(),
-        ]);
+        $session = Yii::$app->session;
+        if ($session->get('user')) {
+          if ($session->get('user')->scope == "admin") {
+            $dataProvider = new ActiveDataProvider([
+                'query' => Test::find(),
+            ]);
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'dataProvider' => $dataProvider,
+            ]);
+          }
+        }
+        return $this->goBack();
     }
 
     /**
@@ -48,9 +54,15 @@ class TestdataController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        $session = Yii::$app->session;
+        if ($session->get('user')) {
+          if ($session->get('user')->scope == "admin") {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+          }
+        }
+        return $this->goBack();
     }
 
     /**
@@ -60,15 +72,20 @@ class TestdataController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Test();
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+        $session = Yii::$app->session;
+        if ($session->get('user')) {
+          if ($session->get('user')->scope == "admin") {
+            $model = new Test();
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('create', [
+                    'model' => $model,
+                ]);
+            }
+          }
         }
+        return $this->redirect(Yii::$app->homeUrl);
     }
 
     /**
@@ -79,15 +96,21 @@ class TestdataController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        $session = Yii::$app->session;
+        if ($session->get('user')) {
+          if ($session->get('user')->scope == "admin") {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
+          }
         }
+        return $this->goBack();
     }
 
     /**
@@ -98,9 +121,15 @@ class TestdataController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        $session = Yii::$app->session;
+        if ($session->get('user')) {
+          if ($session->get('user')->scope == "admin") {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+          }
+        }
+        return $this->goBack();
     }
 
     /**
@@ -112,10 +141,16 @@ class TestdataController extends Controller
      */
     protected function findModel($id)
     {
-        if (($model = Test::findOne($id)) !== null) {
-            return $model;
-        } else {
-            throw new NotFoundHttpException('The requested page does not exist.');
+        $session = Yii::$app->session;
+        if ($session->get('user')) {
+          if ($session->get('user')->scope == "admin") {
+            if (($model = Test::findOne($id)) !== null) {
+                return $model;
+            } else {
+                throw new NotFoundHttpException('The requested page does not exist.');
+            }
+          }
         }
+        return $this->goBack();
     }
 }
